+++ rpc/src/client/transport/http.rs
@@ -1,11 +1,13 @@
 //! HTTP-based transport for Tendermint RPC Client.
 
+use async_trait::async_trait;
+use hyper::body::Buf;
+use hyper::header;
+
+use tendermint::net;
+
 use crate::client::transport::utils::get_tcp_host_port;
 use crate::{Client, Response, Result, SimpleRequest};
-use async_trait::async_trait;
-use bytes::buf::ext::BufExt;
-use hyper::header;
-use tendermint::net;
 
 /// A JSON-RPC/HTTP Tendermint RPC client (implements [`Client`]).
 ///
@@ -62,7 +64,8 @@ impl Client for HttpClient {
                     .unwrap(),
             );
         }
-        let http_client = hyper::Client::builder().build_http();
+
+        let http_client = hyper::Client::new();
         let response = http_client.request(request).await?;
         let response_body = hyper::body::aggregate(response.into_body()).await?;
         R::Response::from_reader(response_body.reader())