+++ tools/kvstore-test/tests/light-client.rs
@@ -1,38 +1,47 @@
 //! Light Client integration tests.
 //!
-/// If you have a kvstore app running on 127.0.0.1:26657,
-/// these can be run using:
-///
-///     cargo test
-///
-/// Or else, if you have docker installed, you can tell the tests to run an endpoint,
-/// by running:
-///
-///     cargo make
-///
-/// (Make sure you install cargo-make using `cargo install cargo-make` first.)
-///
+//! If you have a kvstore app running on 127.0.0.1:26657,
+//! these can be run using:
+//!
+//!     cargo test
+//!
+//! Or else, if you have docker installed, you can tell the tests to run an endpoint,
+//! by running:
+//!
+//!     cargo make
+//!
+//! (Make sure you install cargo-make using `cargo install cargo-make` first.)
+//!
 
 use tendermint_light_client::{
-    builder::LightClientBuilder,
-    builder::SupervisorBuilder,
-    components::io::AtHeight,
-    components::io::Io,
-    components::io::IoError,
-    components::io::ProdIo,
+    builder::{LightClientBuilder, SupervisorBuilder},
+    components::io::{AtHeight, Io, IoError, ProdIo},
+    errors::Error,
     evidence::{Evidence, EvidenceReporter},
     light_client,
-    store::memory::MemoryStore,
-    store::LightStore,
-    supervisor::{Handle, Instance},
-    types::{PeerId, Status, TrustThreshold},
+    store::{memory::MemoryStore, LightStore},
+    supervisor::{Handle, Instance, Supervisor},
+    types::{Height, PeerId, Status, TrustThreshold},
 };
 
 use tendermint::abci::transaction::Hash as TxHash;
 use tendermint::net;
 use tendermint_rpc as rpc;
 
+use std::convert::TryFrom;
 use std::time::Duration;
+
+struct TestEvidenceReporter;
+
+#[contracts::contract_trait]
+impl EvidenceReporter for TestEvidenceReporter {
+    fn report(&self, evidence: Evidence, peer: PeerId) -> Result<TxHash, IoError> {
+        panic!(
+            "unexpected fork detected for peer {} with evidence: {:?}",
+            peer, evidence
+        );
+    }
+}
 
 fn make_instance(
     peer_id: PeerId,
@@ -58,20 +67,7 @@ fn make_instance(
     .build()
 }
 
-struct TestEvidenceReporter;
-
-#[contracts::contract_trait]
-impl EvidenceReporter for TestEvidenceReporter {
-    fn report(&self, evidence: Evidence, peer: PeerId) -> Result<TxHash, IoError> {
-        panic!(
-            "unexpected fork detected for peer {} with evidence: {:?}",
-            peer, evidence
-        );
-    }
-}
-
-#[test]
-fn sync() {
+fn make_supervisor() -> Supervisor {
     let primary: PeerId = "BADFADAD0BEFEEDC0C0ADEADBEEFC0FFEEFACADE".parse().unwrap();
     let witness: PeerId = "CEFEEDBADFADAD0C0CEEFACADE0ADEADBEEFC0FF".parse().unwrap();
 
@@ -99,6 +95,13 @@ fn sync() {
         .witness(witness, node_address, witness_instance)
         .build_prod();
 
+    supervisor
+}
+
+#[test]
+fn forward() {
+    let supervisor = make_supervisor();
+
     let handle = supervisor.handle();
     std::thread::spawn(|| supervisor.run());
 
@@ -119,4 +122,36 @@ fn sync() {
 
         std::thread::sleep(Duration::from_millis(800));
     }
+}
+
+#[test]
+fn backward() -> Result<(), Error> {
+    let supervisor = make_supervisor();
+
+    let handle = supervisor.handle();
+    std::thread::spawn(|| supervisor.run());
+
+    let max_iterations: usize = 10;
+
+    // Sleep a little bit to ensure we have a few blocks already
+    std::thread::sleep(Duration::from_secs(2));
+
+    for i in 1..=max_iterations {
+        println!("[info ] - iteration {}/{}", i, max_iterations);
+
+        // First we sync to the highest block to have a high enough trusted state
+        let trusted_state = handle.verify_to_highest()?;
+        println!("[info ] synced to highest block {}", trusted_state.height());
+
+        // Then we pick a height below the trusted state
+        let target_height = Height::try_from(trusted_state.height().value() / 2).unwrap();
+
+        // We now try to verify a block at this height
+        let light_block = handle.verify_to_target(target_height)?;
+        println!("[info ] verified lower block {}", light_block.height());
+
+        std::thread::sleep(Duration::from_millis(800));
+    }
+
+    Ok(())
 }